requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.4
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==4.0.3
attrs==25.3.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
dateparser==1.2.1
deprecation==2.1.0
discord==2.3.2
discord.py==2.5.2
distro==1.9.0
exceptiongroup==1.3.0
frozenlist==1.6.0
google-api-core==2.24.2
google-api-python-client==2.170.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
gotrue==2.12.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.25
langchain-core==0.3.63
langchain-openai==0.3.18
langchain-tests==0.3.19
langchain-text-splitters==0.3.8
langsmith==0.3.43
multidict==6.4.4
numpy==2.2.6
oauthlib==3.2.2
openai==1.82.1
orjson==3.10.18
packaging==24.2
pinecone==6.0.2
pinecone-client==6.0.0
pinecone-plugin-interface==0.0.7
pluggy==1.6.0
postgrest==1.0.2
propcache==0.3.1
proto-plus==1.26.1
protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
PyJWT==2.10.1
pyparsing==3.2.3
pytest==8.3.5
pytest-asyncio==0.26.0
pytest-mock==3.14.1
pytest-socket==0.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
realtime==2.4.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
storage3==0.11.3
StrEnum==0.4.15
supabase==2.15.2
supafunc==0.9.4
syrupy==4.9.1
tenacity==9.1.2
tiktoken==0.9.0
tomli==2.2.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.13.2
tzlocal==5.3.1
uritemplate==4.1.1
urllib3==2.4.0
websockets==14.2
yarl==1.20.0
zstandard==0.23.0
